Home.js is now refactored into a much smaller component
