+++ src/containers/SearchBar.js
@@ -49,6 +49,7 @@ class SearchBar extends Component {
 
   handleSubmit(event) {
     event.preventDefault();
+    this.props.router.history.push('/admin');
     this.props.writeSearchTerm(this.state.searchTerm, this.state.searchCategory);
   }
